Generate index.html files

+++ build.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-REPO=git@github.com:jaburns/static-repo
+REPO=git@github.com:jaburns/ion
 STYLE=autumn
 
 if [[ "$1" == '--inner' ]]; then
@@ -10,17 +10,54 @@ if [[ "$1" == '--inner' ]]; then
 
     cd repo
 
-    find . | while read path; do
+    find . > /tmp/tree
+
+    cat /tmp/tree | while read path; do
         [[ -d "$path" ]] && continue
 
         if pygmentize -f html "$path" >/tmp/html; then
-            cat /tmp/css /tmp/html >"$path".html
+            cat /tmp/css /tmp/html >"$path".html_gen
         else
-            cp "$path" "$path".txt
+            cp "$path" "$path".txt_gen
         fi
-
-        rm "$path"
     done
+
+    cat /tmp/tree | while read path; do
+        [[ -d "$path" ]] || continue
+
+        echo '<ul><li><a href="..">..</a></li>' >"$path"/index.html_gen
+
+        find "$path" -maxdepth 1 | grep -v '_gen$' > /tmp/inner
+
+        tail -n+2 /tmp/inner | while read item; do
+            item_show="$(printf $item | sed 's/^\.//')"
+
+            if [[ -d "$item" ]]; then
+                item_href="$item_show"
+            elif [[ -f "$item".html_gen ]]; then
+                item_href="$item_show".html
+            else
+                item_href="$item_show".txt
+            fi
+
+            echo "<li><a href=\"$item_href\">$item_show</a></li>" >>"$path"/index.html_gen
+        done
+
+        echo '</ul>' >>"$path"/index.html_gen
+    done
+
+    find . > /tmp/tree
+
+    cat /tmp/tree | while read path; do
+        [[ -d "$path" ]] && continue
+
+        if [[ "$path" == *_gen ]]; then
+            mv "$path" "$(printf "$path" | sed 's/_gen$//')"
+        else
+            rm "$path"
+        fi
+    done
+
 else
     [[ ! -d repo ]] && git clone "$REPO" repo
     docker build -t static-repo .
